sach

+++ src/models/TelegramUser.js
@@ -7,7 +7,7 @@ const TelegramUser = sequelize.define("TelegramUser", {
     tname: { type: DataTypes.STRING, allowNull: true }, // Stores last trade timestamp
     tlastname: { type: DataTypes.STRING, allowNull: true }, // Stores last trade timestamp
     balance: { type: DataTypes.FLOAT, allowNull: true }, // Stores last trade timestamp
-    tabbalance: { type: DataTypes.INTEGER, allowNull: true }, // Stores last trade timestamp
+    tabbalance: { type: DataTypes.FLOAT, allowNull: true }, // Stores last trade timestamp
     invite_bonus: {type: DataTypes.FLOAT, allowNull: false,},// Stores last trade timestam
     sponsor: {type: DataTypes.STRING, allowNull: true,},    
 }, {
